gitconfig: revert color changes
I moved to using color names in commit 8cf70bb4 after upgrading to F28
because of the lack of bold fonts, which I thought was by design :)
Turns out, it was a bug (BZ#1575649), which is now fixed and so the
"bold" variants are now really bold, again.

Therefore, turn back to color codes.